Re-organize stable and unstable txpool options
