+++ besu/src/main/java/org/hyperledger/besu/cli/DefaultCommandValues.java
@@ -16,7 +16,6 @@ package org.hyperledger.besu.cli;
 
 import org.hyperledger.besu.datatypes.Wei;
 import org.hyperledger.besu.ethereum.api.jsonrpc.authentication.JwtAlgorithm;
-import org.hyperledger.besu.ethereum.eth.transactions.TransactionPoolConfiguration;
 import org.hyperledger.besu.ethereum.p2p.config.RlpxConfiguration;
 import org.hyperledger.besu.nat.NatMethod;
 
@@ -61,9 +60,6 @@ public interface DefaultCommandValues {
   String MANDATORY_NODE_ID_FORMAT_HELP = "<NODEID>";
   /** The constant DEFAULT_MIN_TRANSACTION_GAS_PRICE. */
   Wei DEFAULT_MIN_TRANSACTION_GAS_PRICE = Wei.of(1000);
-  /** The constant DEFAULT_RPC_TX_FEE_CAP. */
-  Wei DEFAULT_RPC_TX_FEE_CAP = TransactionPoolConfiguration.DEFAULT_RPC_TX_FEE_CAP;
-
   /** The constant DEFAULT_MIN_BLOCK_OCCUPANCY_RATIO. */
   Double DEFAULT_MIN_BLOCK_OCCUPANCY_RATIO = 0.8;
   /** The constant DEFAULT_EXTRA_DATA. */
