+++ besu/src/main/java/org/hyperledger/besu/cli/options/unstable/TransactionPoolOptions.java
@@ -14,8 +14,6 @@
  */
 package org.hyperledger.besu.cli.options.unstable;
 
-import static org.hyperledger.besu.ethereum.eth.transactions.TransactionPoolConfiguration.Implementation.LAYERED;
-
 import org.hyperledger.besu.cli.options.CLIOptions;
 import org.hyperledger.besu.cli.options.OptionParser;
 import org.hyperledger.besu.ethereum.eth.transactions.ImmutableTransactionPoolConfiguration;
@@ -25,39 +23,15 @@ import java.time.Duration;
 import java.util.Arrays;
 import java.util.List;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
 
 /** The Transaction pool Cli options. */
-public class TransactionPoolOptions
-    implements CLIOptions<ImmutableTransactionPoolConfiguration.Builder> {
-  private static final Logger LOG = LoggerFactory.getLogger(TransactionPoolOptions.class);
-
+public class TransactionPoolOptions implements CLIOptions<TransactionPoolConfiguration> {
   private static final String TX_MESSAGE_KEEP_ALIVE_SEC_FLAG =
       "--Xincoming-tx-messages-keep-alive-seconds";
 
   private static final String ETH65_TX_ANNOUNCED_BUFFERING_PERIOD_FLAG =
       "--Xeth65-tx-announced-buffering-period-milliseconds";
-
-  private static final String STRICT_TX_REPLAY_PROTECTION_ENABLED_FLAG =
-      "--strict-tx-replay-protection-enabled";
-
-  private static final String TX_POOL_IMPLEMENTATION = "--tx-pool";
-  private static final String LAYERED_TX_POOL_LAYER_MAX_CAPACITY =
-      "--layered-tx-pool-layer-max-capacity";
-  private static final String LAYERED_TX_POOL_MAX_PRIORITIZED = "--layered-tx-pool-max-prioritized";
-  private static final String LAYERED_TX_POOL_MAX_FUTURE_BY_SENDER =
-      "--layered-tx-pool-max-future-by-sender";
-
-  @CommandLine.Option(
-      names = {STRICT_TX_REPLAY_PROTECTION_ENABLED_FLAG},
-      paramLabel = "<Boolean>",
-      description =
-          "Require transactions submitted via JSON-RPC to use replay protection in accordance with EIP-155 (default: ${DEFAULT-VALUE})",
-      fallbackValue = "true",
-      arity = "0..1")
-  private Boolean strictTxReplayProtectionEnabled = false;
 
   @CommandLine.Option(
       names = {TX_MESSAGE_KEEP_ALIVE_SEC_FLAG},
@@ -78,43 +52,6 @@ public class TransactionPoolOptions
       arity = "1")
   private long eth65TrxAnnouncedBufferingPeriod =
       TransactionPoolConfiguration.ETH65_TRX_ANNOUNCED_BUFFERING_PERIOD.toMillis();
-
-  @CommandLine.Option(
-      names = {TX_POOL_IMPLEMENTATION},
-      paramLabel = "<Enum>",
-      description = "The Transaction Pool implementation to use(default: ${DEFAULT-VALUE})",
-      arity = "0..1")
-  private TransactionPoolConfiguration.Implementation txPoolImplementation = LAYERED;
-
-  @CommandLine.Option(
-      names = {LAYERED_TX_POOL_LAYER_MAX_CAPACITY},
-      paramLabel = "<Long>",
-      hidden = true,
-      description =
-          "Max amount of memory space, in bytes, that any layer within the transaction pool could occupy (default: ${DEFAULT-VALUE})",
-      arity = "1")
-  private long layeredTxPoolLayerMaxCapacity =
-      TransactionPoolConfiguration.DEFAULT_PENDING_TRANSACTIONS_LAYER_MAX_CAPACITY_BYTES;
-
-  @CommandLine.Option(
-      names = {LAYERED_TX_POOL_MAX_PRIORITIZED},
-      paramLabel = "<Int>",
-      hidden = true,
-      description =
-          "Max number of pending transactions that are prioritized and thus kept sorted (default: ${DEFAULT-VALUE})",
-      arity = "1")
-  private int layeredTxPoolMaxPrioritized =
-      TransactionPoolConfiguration.DEFAULT_MAX_PRIORITIZED_TRANSACTIONS;
-
-  @CommandLine.Option(
-      names = {LAYERED_TX_POOL_MAX_FUTURE_BY_SENDER},
-      paramLabel = "<Int>",
-      hidden = true,
-      description =
-          "Max number of future pending transactions allowed for a single sender (default: ${DEFAULT-VALUE})",
-      arity = "1")
-  private int layeredTxPoolMaxFutureBySender =
-      TransactionPoolConfiguration.DEFAULT_MAX_FUTURE_BY_SENDER;
 
   private TransactionPoolOptions() {}
 
@@ -138,46 +75,23 @@ public class TransactionPoolOptions
     options.txMessageKeepAliveSeconds = config.getTxMessageKeepAliveSeconds();
     options.eth65TrxAnnouncedBufferingPeriod =
         config.getEth65TrxAnnouncedBufferingPeriod().toMillis();
-    options.strictTxReplayProtectionEnabled = config.getStrictTransactionReplayProtectionEnabled();
-    options.txPoolImplementation = config.getTxPoolImplementation();
-    options.layeredTxPoolLayerMaxCapacity = config.getPendingTransactionsLayerMaxCapacityBytes();
-    options.layeredTxPoolMaxPrioritized = config.getMaxPrioritizedTransactions();
-    options.layeredTxPoolMaxFutureBySender = config.getMaxFutureBySender();
     return options;
   }
 
   @Override
-  public ImmutableTransactionPoolConfiguration.Builder toDomainObject() {
-    if (txPoolImplementation.equals(LAYERED)) {
-      LOG.warn(
-          "Layered transaction pool enabled, ignoring settings for "
-              + "--tx-pool-max-size and --tx-pool-limit-by-account-percentage");
-    }
-
+  public TransactionPoolConfiguration toDomainObject() {
     return ImmutableTransactionPoolConfiguration.builder()
-        .strictTransactionReplayProtectionEnabled(strictTxReplayProtectionEnabled)
         .txMessageKeepAliveSeconds(txMessageKeepAliveSeconds)
         .eth65TrxAnnouncedBufferingPeriod(Duration.ofMillis(eth65TrxAnnouncedBufferingPeriod))
-        .txPoolImplementation(txPoolImplementation)
-        .pendingTransactionsLayerMaxCapacityBytes(layeredTxPoolLayerMaxCapacity)
-        .maxPrioritizedTransactions(layeredTxPoolMaxPrioritized)
-        .maxFutureBySender(layeredTxPoolMaxFutureBySender);
+        .build();
   }
 
   @Override
   public List<String> getCLIOptions() {
     return Arrays.asList(
-        STRICT_TX_REPLAY_PROTECTION_ENABLED_FLAG + "=" + strictTxReplayProtectionEnabled,
         TX_MESSAGE_KEEP_ALIVE_SEC_FLAG,
         OptionParser.format(txMessageKeepAliveSeconds),
         ETH65_TX_ANNOUNCED_BUFFERING_PERIOD_FLAG,
-        OptionParser.format(eth65TrxAnnouncedBufferingPeriod),
-        TX_POOL_IMPLEMENTATION + "=" + txPoolImplementation,
-        LAYERED_TX_POOL_LAYER_MAX_CAPACITY,
-        OptionParser.format(layeredTxPoolLayerMaxCapacity),
-        LAYERED_TX_POOL_MAX_PRIORITIZED,
-        OptionParser.format(layeredTxPoolMaxPrioritized),
-        LAYERED_TX_POOL_MAX_FUTURE_BY_SENDER,
-        OptionParser.format(layeredTxPoolMaxFutureBySender));
+        OptionParser.format(eth65TrxAnnouncedBufferingPeriod));
   }
 }
