+++ besu/src/test/java/org/hyperledger/besu/cli/options/MetricsCLIOptionsTest.java
@@ -25,24 +25,25 @@ public class MetricsCLIOptionsTest
     extends AbstractCLIOptionsTest<MetricsConfiguration.Builder, MetricsCLIOptions> {
 
   @Override
-  MetricsConfiguration.Builder createDefaultDomainObject() {
+  protected MetricsConfiguration.Builder createDefaultDomainObject() {
     return MetricsConfiguration.builder();
   }
 
   @Override
-  MetricsConfiguration.Builder createCustomizedDomainObject() {
+  protected MetricsConfiguration.Builder createCustomizedDomainObject() {
     return MetricsConfiguration.builder()
         .timersEnabled(!MetricsConfiguration.DEFAULT_METRICS_TIMERS_ENABLED)
         .idleTimeout(MetricsConfiguration.DEFAULT_METRICS_IDLE_TIMEOUT_SECONDS);
   }
 
   @Override
-  MetricsCLIOptions optionsFromDomainObject(final MetricsConfiguration.Builder domainObject) {
+  protected MetricsCLIOptions optionsFromDomainObject(
+      final MetricsConfiguration.Builder domainObject) {
     return MetricsCLIOptions.fromConfiguration(domainObject.build());
   }
 
   @Override
-  MetricsCLIOptions getOptionsFromBesuCommand(final TestBesuCommand besuCommand) {
+  protected MetricsCLIOptions getOptionsFromBesuCommand(final TestBesuCommand besuCommand) {
     return besuCommand.getMetricsCLIOptions();
   }
 }
