+++ besu/src/test/java/org/hyperledger/besu/cli/options/stable/TransactionPoolOptionsTest.java
@@ -12,17 +12,15 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.cli.options;
+package org.hyperledger.besu.cli.options.stable;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hyperledger.besu.ethereum.eth.transactions.TransactionPoolConfiguration.Implementation.LAYERED;
 
-import org.hyperledger.besu.cli.options.unstable.TransactionPoolOptions;
+import org.hyperledger.besu.cli.options.AbstractCLIOptionsTest;
 import org.hyperledger.besu.ethereum.eth.transactions.ImmutableTransactionPoolConfiguration;
 import org.hyperledger.besu.ethereum.eth.transactions.TransactionPoolConfiguration;
-
-import java.time.Duration;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -30,15 +28,14 @@ import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
 public class TransactionPoolOptionsTest
-    extends AbstractCLIOptionsTest<
-        ImmutableTransactionPoolConfiguration.Builder, TransactionPoolOptions> {
+    extends AbstractCLIOptionsTest<TransactionPoolConfiguration, TransactionPoolOptions> {
 
   @Test
   public void strictTxReplayProtection_enabled() {
     final TestBesuCommand cmd = parseCommand("--strict-tx-replay-protection-enabled");
 
     final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
-    final TransactionPoolConfiguration config = options.toDomainObject().build();
+    final TransactionPoolConfiguration config = options.toDomainObject();
     assertThat(config.getStrictTransactionReplayProtectionEnabled()).isTrue();
 
     assertThat(commandOutput.toString(UTF_8)).isEmpty();
@@ -50,7 +47,7 @@ public class TransactionPoolOptionsTest
     final TestBesuCommand cmd = parseCommand("--strict-tx-replay-protection-enabled=true");
 
     final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
-    final TransactionPoolConfiguration config = options.toDomainObject().build();
+    final TransactionPoolConfiguration config = options.toDomainObject();
     assertThat(config.getStrictTransactionReplayProtectionEnabled()).isTrue();
 
     assertThat(commandOutput.toString(UTF_8)).isEmpty();
@@ -62,7 +59,7 @@ public class TransactionPoolOptionsTest
     final TestBesuCommand cmd = parseCommand("--strict-tx-replay-protection-enabled=false");
 
     final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
-    final TransactionPoolConfiguration config = options.toDomainObject().build();
+    final TransactionPoolConfiguration config = options.toDomainObject();
     assertThat(config.getStrictTransactionReplayProtectionEnabled()).isFalse();
 
     assertThat(commandOutput.toString(UTF_8)).isEmpty();
@@ -74,83 +71,47 @@ public class TransactionPoolOptionsTest
     final TestBesuCommand cmd = parseCommand();
 
     final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
-    final TransactionPoolConfiguration config = options.toDomainObject().build();
+    final TransactionPoolConfiguration config = options.toDomainObject();
     assertThat(config.getStrictTransactionReplayProtectionEnabled()).isFalse();
 
     assertThat(commandOutput.toString(UTF_8)).isEmpty();
     assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
   }
 
-  @Test
-  public void txMessageKeepAliveSeconds() {
-    final int txMessageKeepAliveSeconds = 999;
-    final TestBesuCommand cmd =
-        parseCommand(
-            "--Xincoming-tx-messages-keep-alive-seconds",
-            String.valueOf(txMessageKeepAliveSeconds));
-
-    final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
-    final TransactionPoolConfiguration config = options.toDomainObject().build();
-    assertThat(config.getTxMessageKeepAliveSeconds()).isEqualTo(txMessageKeepAliveSeconds);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void eth65TrxAnnouncedBufferingPeriod() {
-    final long eth65TrxAnnouncedBufferingPeriod = 999;
-    final TestBesuCommand cmd =
-        parseCommand(
-            "--Xeth65-tx-announced-buffering-period-milliseconds",
-            String.valueOf(eth65TrxAnnouncedBufferingPeriod));
-
-    final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
-    final TransactionPoolConfiguration config = options.toDomainObject().build();
-    assertThat(config.getEth65TrxAnnouncedBufferingPeriod())
-        .hasMillis(eth65TrxAnnouncedBufferingPeriod);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
   @Override
-  ImmutableTransactionPoolConfiguration.Builder createDefaultDomainObject() {
+  protected TransactionPoolConfiguration createDefaultDomainObject() {
     final ImmutableTransactionPoolConfiguration defaultValue =
         ImmutableTransactionPoolConfiguration.builder().build();
     return ImmutableTransactionPoolConfiguration.builder()
+        .from(defaultValue)
         .strictTransactionReplayProtectionEnabled(false)
-        .txMessageKeepAliveSeconds(defaultValue.getTxMessageKeepAliveSeconds())
-        .eth65TrxAnnouncedBufferingPeriod(defaultValue.getEth65TrxAnnouncedBufferingPeriod())
         .txPoolImplementation(defaultValue.getTxPoolImplementation())
         .pendingTransactionsLayerMaxCapacityBytes(
             defaultValue.getPendingTransactionsLayerMaxCapacityBytes())
         .maxPrioritizedTransactions(defaultValue.getMaxPrioritizedTransactions())
-        .maxFutureBySender(defaultValue.getMaxFutureBySender());
+        .maxFutureBySender(defaultValue.getMaxFutureBySender())
+        .build();
   }
 
   @Override
-  ImmutableTransactionPoolConfiguration.Builder createCustomizedDomainObject() {
+  protected TransactionPoolConfiguration createCustomizedDomainObject() {
     return ImmutableTransactionPoolConfiguration.builder()
         .strictTransactionReplayProtectionEnabled(true)
-        .txMessageKeepAliveSeconds(TransactionPoolConfiguration.DEFAULT_TX_MSG_KEEP_ALIVE + 1)
-        .eth65TrxAnnouncedBufferingPeriod(
-            TransactionPoolConfiguration.ETH65_TRX_ANNOUNCED_BUFFERING_PERIOD.plus(
-                Duration.ofMillis(100)))
         .txPoolImplementation(LAYERED)
         .pendingTransactionsLayerMaxCapacityBytes(1_000_000L)
         .maxPrioritizedTransactions(1000)
-        .maxFutureBySender(10);
+        .maxFutureBySender(10)
+        .build();
   }
 
   @Override
-  TransactionPoolOptions optionsFromDomainObject(
-      final ImmutableTransactionPoolConfiguration.Builder domainObject) {
-    return TransactionPoolOptions.fromConfig(domainObject.build());
+  protected TransactionPoolOptions optionsFromDomainObject(
+      final TransactionPoolConfiguration domainObject) {
+    return TransactionPoolOptions.fromConfig(domainObject);
   }
 
   @Override
-  TransactionPoolOptions getOptionsFromBesuCommand(final TestBesuCommand besuCommand) {
-    return besuCommand.getTransactionPoolOptions();
+  protected TransactionPoolOptions getOptionsFromBesuCommand(final TestBesuCommand besuCommand) {
+    return besuCommand.getStableTransactionPoolOptions();
   }
 }
